refactor: create RefItemRendererComp and RenderFoundItemComp components for improved Bible reference rendering
